Moved dataflow utils to accessor and creates types.go
